refactor: fixing up files and directories
- moving the files to a new directory, relevant to your goal;
- also fixing the imports inside the files.
