+++ moneymind/src/insert/get_and_insert_institutions.py
@@ -1,5 +1,5 @@
-from get_institutions import get_institutions
-from mysql_initializer import establishing_mysql_connection
+from src.get.get_institutions import get_institutions
+from src.mysql_initializer import establishing_mysql_connection
 
 connection = establishing_mysql_connection()
 cursor = connection.cursor()
