+++ moneymind/src/main.py
@@ -1,10 +1,10 @@
 from mysql_initializer import establishing_mysql_connection
 
-from get_and_insert_all_tariffs_and_value import get_and_insert_all_tariffs_and_value
-from get_and_insert_institutions import get_and_insert_institutions
-from get_and_insert_institutions_tariffs import get_and_insert_institutions_tariffs
-from insert_all_data_code_services import insert_all_data_code_services
-from insert_consolidated_group_data import insert_consolidated_group_data
+from src.insert.get_and_insert_all_tariffs_and_value import get_and_insert_all_tariffs_and_value
+from src.insert.get_and_insert_institutions import get_and_insert_institutions
+from src.insert.get_and_insert_institutions_tariffs import get_and_insert_institutions_tariffs
+from src.insert.insert_all_data_code_services import insert_all_data_code_services
+from src.insert.insert_consolidated_group_data import insert_consolidated_group_data
 
 connection = establishing_mysql_connection()
 cursor = connection.cursor()
@@ -16,6 +16,7 @@ def application_startup():
     get_and_insert_all_tariffs_and_value()
     get_and_insert_institutions()
     get_and_insert_institutions_tariffs()
+
 
 application_startup()
 cursor.close()
